requirements.txt
streamlit>=1.30
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=0.27.0
python-dotenv>=1.0.0
requests>=2.32.0